major update with several extensions and fixes; version number bumbed to 1.1.9
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/kebabs@102273 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/GappyPairC.cpp
@@ -5,7 +5,7 @@
 // Package: kebabs
 // Author : J. P.
 //
-// Copyright (C) 2014   J o h a n n e s  P a l m e
+// Copyright (C) 2014-2015   J o h a n n e s  P a l m e
 //
 
 #include "Kebabs.h"
@@ -929,7 +929,7 @@ void getKMStdAnnGappy(T maxUnSignedIndex, NumericMatrix km, ByteStringVector x, 
         noSamples += sizeY;
 
     // add one for the sentinel
-    if (dimFeatureSpace < maxSeqLength)
+    if (dimFeatureSpace < (uint64_t) maxSeqLength)
         maxNumFeatures = dimFeatureSpace + 1;
     else
         maxNumFeatures = (maxSeqLength - 2*k - 0.5*m + 1) * (m + 1) + 1;
@@ -1552,7 +1552,7 @@ void getKMPosDistGappy(T maxUnSignedIndex, NumericMatrix km, ByteStringVector x,
 
         featVectorsStart[i + 1] = elemIndex;
 
-        if (maxFeaturesPerSample < (featVectorsStart[i + 1] - featVectorsStart[i]))
+        if (maxFeaturesPerSample < (int) (featVectorsStart[i + 1] - featVectorsStart[i]))
             maxFeaturesPerSample = featVectorsStart[i + 1] - featVectorsStart[i];
 
         // for dist weighting the kernel value is determined in computeKernelMatrixPos
